Fixes #1461

Adds an InPlace median finding to improve the performance of the metric aggregates.
We used to sort a sequence to create StatisticsMetrics which turned out to be very expensive in large eventlogs.

It was found that there is a bottleneck in `generateStageLevelAccums` and `generateSQLAccums` due to the cost of sorting and using sequences of Long on metrics aggregates.

### Impact on performance:

The PR improves the Qualification runtime. methods like

- `generateSQLAccums` is down from 21,550 to 3,480 ms (~80% improvement)
- `generateStageLevelAccums`  is down from  115580 to 61,611 ms (~45% improvement)

### Main optimization

_Median finding_: we used to convert a map to sequence. Then get it sorted. Then we pick the median, max, and min, and then we call `seq.sum`
  - This turned to be expensive for large eventlogs
  - This PR implements adds finding-median in linear time. On average, this is a linear complexity compared to sorting
- This pull request includes several changes to the `AppSQLPlanAnalyzer` and `AppSparkMetricsAnalyzer` classes to improve performance and simplify the codebase by using the `breakOut` method for collection transformations and introducing new methods in `StatisticsMetrics`. Additionally, it removes an unused object and refactors the `AccumInfo` class.

### Impact on output

By doing a diff on the output folder, it was found that the `sql_plan_metrics_for_application.csv` is differemt. The new output sounds more correct

In output generated by the dev branch:

```
appIndex,sqlID,nodeID,nodeName,accumulatorId,name,min,median,max,total,metricType,stageIds
1,236,0,"Execute InsertIntoHadoopFsRelationCommand",0,"number of written files",0,0,0,2,"sum","2"
```
The above seems incorrect because the total is non-zero while min, max and median are zeros.

Vs the new output
```
appIndex,sqlID,nodeID,nodeName,accumulatorId,name,min,median,max,total,metricType,stageIds
1,236,0,"Execute InsertIntoHadoopFsRelationCommand",0,"number of written files",2,2,2,2,"sum","2"
```

### Improvements to Collection Transformations:

* [`core/src/main/scala/com/nvidia/spark/rapids/tool/analysis/AppSQLPlanAnalyzer.scala`](diffhunk://#diff-d56ed990ec181a2ef118996b608fdff3b53818cf6138baba59b51e4eb407d79eL268-R271): Replaced intermediate collection variables with direct transformations using `breakOut` to improve performance. [[1]](diffhunk://#diff-d56ed990ec181a2ef118996b608fdff3b53818cf6138baba59b51e4eb407d79eL268-R271) [[2]](diffhunk://#diff-d56ed990ec181a2ef118996b608fdff3b53818cf6138baba59b51e4eb407d79eL286-R289) [[3]](diffhunk://#diff-d56ed990ec181a2ef118996b608fdff3b53818cf6138baba59b51e4eb407d79eL297-R303) [[4]](diffhunk://#diff-d56ed990ec181a2ef118996b608fdff3b53818cf6138baba59b51e4eb407d79eL328-R321) [[5]](diffhunk://#diff-d56ed990ec181a2ef118996b608fdff3b53818cf6138baba59b51e4eb407d79eL344-L377)
* [`core/src/main/scala/com/nvidia/spark/rapids/tool/analysis/AppSparkMetricsAnalyzer.scala`](diffhunk://#diff-4b0aab10a86746bb7480cc3bde4e013c04707758c61782934c07604443160b40L82-R83): Applied `breakOut` to streamline collection processing and reduce memory overhead. [[1]](diffhunk://#diff-4b0aab10a86746bb7480cc3bde4e013c04707758c61782934c07604443160b40L82-R83) [[2]](diffhunk://#diff-4b0aab10a86746bb7480cc3bde4e013c04707758c61782934c07604443160b40L129-R130) [[3]](diffhunk://#diff-4b0aab10a86746bb7480cc3bde4e013c04707758c61782934c07604443160b40L166-R166) [[4]](diffhunk://#diff-4b0aab10a86746bb7480cc3bde4e013c04707758c61782934c07604443160b40L175-R175) [[5]](diffhunk://#diff-4b0aab10a86746bb7480cc3bde4e013c04707758c61782934c07604443160b40L232-R232) [[6]](diffhunk://#diff-4b0aab10a86746bb7480cc3bde4e013c04707758c61782934c07604443160b40L244-R243) [[7]](diffhunk://#diff-4b0aab10a86746bb7480cc3bde4e013c04707758c61782934c07604443160b40L256-R255) [[8]](diffhunk://#diff-4b0aab10a86746bb7480cc3bde4e013c04707758c61782934c07604443160b40L292-R290) [[9]](diffhunk://#diff-4b0aab10a86746bb7480cc3bde4e013c04707758c61782934c07604443160b40L341-R338) [[10]](diffhunk://#diff-4b0aab10a86746bb7480cc3bde4e013c04707758c61782934c07604443160b40L362-R359)

### Introduction of New Methods in `StatisticsMetrics`:

* [`core/src/main/scala/com/nvidia/spark/rapids/tool/analysis/StatisticsMetrics.scala`](diffhunk://#diff-13353883e8b8763b970c1ad85d6c670336d17e492997f1cfb812902b2c436a6cR19-R54): Added `createFromArr` and `createOptionalFromArr` methods to compute statistics directly from arrays, improving code reuse and readability.

### Refactoring and Cleanup:

* [`core/src/main/scala/org/apache/spark/sql/rapids/tool/store/AccumInfo.scala`](diffhunk://#diff-2cdf5cec29c5cfc15962382b2134c8e88b6623afdfd7cc6a81ec3dfc5663b4a1L101-R102): Refactored the `calculateAccStats` method to use the new `createFromArr` method from `StatisticsMetrics`, simplifying the code.
* [`core/src/main/scala/com/nvidia/spark/rapids/tool/analysis/AppSparkMetricsAnalyzer.scala`](diffhunk://#diff-4b0aab10a86746bb7480cc3bde4e013c04707758c61782934c07604443160b40L459-L479): Removed the unused `AppSparkMetricsAnalyzer` object and its `getStatistics` method.